Admin can change user role
Admin now have the ability to change the role of a user. Updated the Roles Controller to create a PUT action for updating a role. Connected role name to a user through the User DTO, so the role associated with the user will appear on the user table. This is where the Admin can update the users role. Created a separate user form as the functionality is a bit different than a standard form.